Update IST-555-Final-v1.3.nlogo
Figured we have the drone change color once it has picked up trash so we know it is carrying something. Then once it has dropped it off it returns to a default base color. I just choose some random purple to make it noticeable.  No worries if you dont think we need this change. just a thought.